Separate common slice to modal slice and booksList slice
