+++ src/components/Application/index.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import styles from "./index.module.css";
 import Main from "components/Main";
 import { Provider } from "react-redux";
-import store from "store";
+import store from "models";
 import { BrowserRouter } from "react-router-dom";
 
 function App() {
