+++ src/models/index.js
@@ -1,10 +1,12 @@
 import { combineReducers, applyMiddleware, compose, createStore } from "redux";
 import thunk from "redux-thunk";
-import booksReducer from "store/slice";
+import booksReducer from "models/booksList/slice";
+import modalReducer from "models/modal/slice";
 import { api } from "agent";
 
 const reducers = combineReducers({
   books: booksReducer.reducer,
+  modal: modalReducer.reducer,
 });
 
 const composedEnhancers = compose(
